requirements.txt
flask
python-docx
PyPDF2
cohere
python-dotenv

